[Glimmer] improve curely/text spacing
