+++ tests/glimmer/curly.hbs
@@ -1,2 +1,11 @@
 <p>Your username is @{{name}}</p>
 <p>Hi {{firstName}} {{lastName}}</p>
+
+<div>
+  {{name}}
+  is your name
+</div>
+
+<p>
+  {{alex}} Ben {{sophie}}
+</p>
